Extracted folder of the Zip of the Portable version deleted
Erroneously, the extracted folder of the zip of the Portable version was
also included in the repository. Only the zip file is needed. The
extracted folder is deleted. (Note : You may find the previous commit
looks like a "lot of changes" because of the presence of the extracted
folder earlier.